refactor: Move message types from standard to own module
